# Optimize coalesce kernel for `StringViewArray` (5-10%)