@alamb 2025-06-06
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing alamb/optimize_gc_string_view (fe2304ac41a68cfcc6d3d3a0e5ad74d72dcd1d2b) to 026356be8f2291f6985b45217ba681a4b3d3f204 [diff](https://github.com/apache/arrow-rs/compare/026356be8f2291f6985b45217ba681a4b3d3f204..fe2304ac41a68cfcc6d3d3a0e5ad74d72dcd1d2b)
BENCH_NAME=concatenate_kernel
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench concatenate_kernel 
BENCH_FILTER=
BENCH_BRANCH_NAME=alamb_optimize_gc_string_view
Results will be posted here when complete