@alamb 2025-06-06
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                   alamb_optimize_gc_string_view          main
-----                                                   -----------------------------          ----
concat 1024 arrays boolean 4                            1.00     28.8±0.12µs        ? ?/sec    1.01     29.0±0.05µs        ? ?/sec
concat 1024 arrays i32 4                                1.00     15.4±0.02µs        ? ?/sec    1.06     16.4±0.05µs        ? ?/sec
concat 1024 arrays str 4                                1.02     57.1±0.54µs        ? ?/sec    1.00     55.8±0.43µs        ? ?/sec
concat boolean 1024                                     1.00    336.6±0.70ns        ? ?/sec    1.30    438.4±0.78ns        ? ?/sec
concat boolean 8192 over 100 arrays                     1.00     44.1±0.05µs        ? ?/sec    1.00     44.2±0.05µs        ? ?/sec
concat boolean nulls 1024                               1.00    651.5±1.25ns        ? ?/sec    1.17    763.2±2.65ns        ? ?/sec
concat boolean nulls 8192 over 100 arrays               1.00     96.1±0.16µs        ? ?/sec    1.00     96.0±0.15µs        ? ?/sec
concat fixed size lists                                 1.12   802.1±44.65µs        ? ?/sec    1.00   717.9±29.00µs        ? ?/sec
concat i32 1024                                         1.00    442.7±0.64ns        ? ?/sec    1.00    443.5±0.68ns        ? ?/sec
concat i32 8192 over 100 arrays                         1.00    208.7±1.20µs        ? ?/sec    1.01    211.4±4.55µs        ? ?/sec
concat i32 nulls 1024                                   1.00    651.4±1.85ns        ? ?/sec    1.14    742.6±1.91ns        ? ?/sec
concat i32 nulls 8192 over 100 arrays                   1.01    270.0±6.21µs        ? ?/sec    1.00    268.6±6.25µs        ? ?/sec
concat str 1024                                         1.03     13.4±1.16µs        ? ?/sec    1.00     12.9±0.75µs        ? ?/sec
concat str 8192 over 100 arrays                         1.02    104.0±1.08ms        ? ?/sec    1.00    102.3±0.88ms        ? ?/sec
concat str nulls 1024                                   1.00      6.7±0.89µs        ? ?/sec    1.00      6.7±0.83µs        ? ?/sec
concat str nulls 8192 over 100 arrays                   1.00     52.6±0.45ms        ? ?/sec    1.00     52.4±0.39ms        ? ?/sec
concat str_dict 1024                                    1.00      2.7±0.01µs        ? ?/sec    1.06      2.9±0.01µs        ? ?/sec
concat str_dict_sparse 1024                             1.01      7.0±0.02µs        ? ?/sec    1.00      6.9±0.03µs        ? ?/sec
concat struct with int32 and dicts size=1024 count=2    1.00      6.6±0.05µs        ? ?/sec    1.03      6.8±0.05µs        ? ?/sec
```


</p>
</details>