@alamb 2025-06-06
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing alamb/optimize_gc_string_view (a9ac1bc41080b75398b19eb776f15ee6f04eb0aa) to 026356be8f2291f6985b45217ba681a4b3d3f204 [diff](https://github.com/apache/arrow-rs/compare/026356be8f2291f6985b45217ba681a4b3d3f204..a9ac1bc41080b75398b19eb776f15ee6f04eb0aa)
BENCH_NAME=coalesce_kernels
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench coalesce_kernels 
BENCH_FILTER=
BENCH_BRANCH_NAME=alamb_optimize_gc_string_view
Results will be posted here when complete